requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.5
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prometheus-client==0.17.0
prometheus-flask-exporter==0.22.4
psutil==5.9.5
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.31.0
typing-extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
